requirements.txt
# Core dependencies - actually used in the current codebase
pandas>=1.5.0
numpy>=1.24.0
torch>=2.0.0

# PyTorch Geometric - core components only
# Note: torch-geometric will pull in necessary sub-dependencies
torch-geometric>=2.4.0

# Data preprocessing
scikit-learn>=1.3.0

# Graph operations
networkx>=3.0

# Visualization libraries
matplotlib>=3.6.0
seaborn>=0.12.0

# Configuration management
pyyaml>=6.0

# Built-in Python libraries (no installation needed):
# - pathlib
# - typing
# - logging
# - pickle
# - json
# - datetime
